+++ m_cheat.c
@@ -53,11 +53,11 @@ static const c_cheat cheat_def[] =
 	{"FPS Counter Ammo",CHEAT_SEQ(KEYD_A,   KEYD_B,     KEYD_L,     KEYD_UP,    KEYD_DOWN,  KEYD_B,         KEYD_LEFT,     KEYD_LEFT),      cheat_fps},
 };
 
-static const uint32_t num_cheats = sizeof(cheat_def) / sizeof (c_cheat);
+static const int16_t num_cheats = sizeof(cheat_def) / sizeof (c_cheat);
 
 static boolean CheckCheats(uint32_t keybuff)
 {
-    for(int32_t i = 0; i < num_cheats; i++)
+    for(int16_t i = 0; i < num_cheats; i++)
     {
         if(cheat_def[i].packed_sequence == keybuff)
         {
